Release Branch section of `Gitflow Workflow`
Explains the purpose of release branches, that they are created from the
develop branch after several features are merged into the develop branch
so that documentation may be added to the branch before the release
branch is officially merged into the master branch and tagged with a
version number.

Covers:

- Starting a release branch
- Finishing a release branch

Added additional image files related to previous commit.