Hotfix Branch section of `Gitflow Workflow`
States the purpose of a hotfix branch, which basically to fix issues
after a release has already been merged into the main branch.

Note that a hotfix must also be merged into the develop branch and the
master branch.  This way we don't have to wait for the next release to
fix the issue.